terraform/functions/cluster_protection/requirements.txt
functions-framework==3.*
google-cloud-container
pyyaml
google-cloud-logging
kubernetes
google-api-python-client